pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "1799LidarLib"
version = "0.0.14"
authors = [
  { name="Owen Flanagan", email="twisterjafla@gmail.com" },
]
description = "Library for interfacing with slamtech rplidars "
readme = "readme.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
  "pyserial",
  "pygame",
  "robotpy",
  "matplotlib",

]


license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/FRC-1799/LidarDetection"
Issues = "https://github.com/FRC-1799/LidarDetection/issues"
[tool.hatch.build.targets.wheel]
packages = ["src/lidarLib"]


